make `rules_nixpkgs_core` independent repository
tl;dr: end users of `rules_nixpkgs` should not be affected by this
change at all. the additional code is temporary and purely for
migration reasons.

migrating the go toolchain to `bzlmod` necessitates accessing
`rules_nixpkgs_core` as an external repository (later module). therefore
we have to incorporate that new repository into the `WORKSPACE` setup.

to maintain compatibility with existing installations,
`rules_nixpkgs_dependencies()` automatically fetches the new
sub-repository from a remote archive. but local testing requires taking
it from the working directory! therefore we temporary introduce
a `local=` parameter which lets the caller point optionally to
a directory for local sub-repositories. the option will be removed when
migration to `bzlmod` is complete.

it must be configurable,  because the call to `local_repository()` takes
its path starting from the workspace directory where
`rules_nixpkgs_dependencies()` is called:
  - `bazel test //...` needs `local="."`
  - `cd examples/toolchains/{go, java, ...}; bazel run :hello` needs
  `local="../../.."`