move generation of core docs into top-level package
remove unused README templates, as rendering is now self-contained.

motivation: `docs` package is for external use by other modules.
modules will have their own documentation rules, and we do not want to
introduce artificial packages in each just for that.

this change requires clearly distinguishing between the transient,
generated file and the source file that we create with `copy_files`.
that was not necessary before, as the transient and source files lived
in different directories. since Bazel does not have a way to express
this distinction, encode the distinction with (arbitrarily) differing
file names.

for now decide against encoding this in the rules, as the mechanism
would be obfuscated.